% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percentile.R
\name{add_percentile}
\alias{add_percentile}
\title{Add percentiles}
\usage{
add_percentile(df, col)
}
\arguments{
\item{df}{A 'data.frame'.}

\item{col}{Column name as a 'string', for which the percentile is calculated.}
}
\value{
A 'data.frame' with additional column '<col>_percentile'.
}
\description{
Adds the percentile (0-100) as a new column to a data frame column
}
\examples{
add_percentile(iris, "Sepal.Length")
}
