% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{calc_mean_sd}
\alias{calc_mean_sd}
\title{Mean and standard deviation}
\usage{
calc_mean_sd(x)
}
\arguments{
\item{x}{Numeric Vector.}
}
\value{
A List with elements 'mean' and 'sd'.
}
\description{
Calculates the mean and standard deviation of a numeric vector column.
}
\examples{
calc_mean_sd(c(1, 2, 3, 4, 5))
}
