% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_missing_values.R
\name{count_missing}
\alias{count_missing}
\title{Count Missing (NA) Values in Each Column}
\usage{
count_missing(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the count of missing values per column.
}
\description{
Count Missing (NA) Values in Each Column
}
\examples{
count_missing(mtcars)
df <- data.frame(a = c(1, NA, 3), b = c(NA, NA, 3))
count_missing(df)
}
