% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter_package.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Create a scatter plot with optional color and shape aesthetics}
\usage{
create_scatter_plot(
  data,
  x_param,
  y_param,
  color_var = NULL,
  shape_var = NULL,
  title = "Scatter Plot",
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{x_param}{The variable for the x-axis.}

\item{y_param}{The variable for the y-axis.}

\item{color_var}{An optional variable to color the points.}

\item{shape_var}{An optional variable to adjust the shape of the points.}

\item{title}{The title of the plot.}

\item{x_label}{A custom label for the x-axis.}

\item{y_label}{A custom label for the y-axis.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function creates a customizable scatter plot using ggplot2.
}
\examples{
data(mtcars)
create_scatter_plot(mtcars, "mpg", "hp", color_var = "cyl", shape_var = "gear")

}
