% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missing_values_heatmap.R
\name{plot_missing_heatmap}
\alias{plot_missing_heatmap}
\title{Plot a Heatmap of Missing Data}
\usage{
plot_missing_heatmap(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A ggplot object showing a heatmap of missingness.
}
\description{
Visualizes the pattern of missing data in a data frame.
}
\examples{
plot_missing_heatmap(airquality)
}
